a9 task 2 fixup: preserve poll_once panic-firewall + tighten elt_size type
Code-quality review flagged that the dispatch_one_rx_mbuf extraction
introduced .expect() on the NonNull::new(mbuf) at the FFI-reachable
poll_once site. Silently skip null slots instead (matches pre-extraction
behaviour; preserves A6.7 panic-firewall for FFI-reachable code).

Also tighten Mempool::elt_size() return type from u32 to u16 to match
the underlying DPDK rte_mempool.elt_size field width and prevent silent
truncation at the shim_rte_pktmbuf_append call site.

+++ crates/dpdk-net-core/src/mempool.rs
@@ -91,11 +91,13 @@ impl Mempool {
     /// Matches the `data_room_size` argument passed to
     /// `rte_pktmbuf_pool_create` — i.e., the maximum single-segment
     /// payload + headroom an mbuf from this pool can hold. Returned as
-    /// `u32` for arithmetic convenience (matches the DPDK C field
-    /// width); callers compare against frame length before copying.
+    /// `u16` to match the underlying DPDK `rte_pktmbuf_pool_private`
+    /// field width (and the `uint16_t` accepted by
+    /// `rte_pktmbuf_append`), which prevents silent truncation at any
+    /// append-style call site that casts the value back to `u16`.
     #[inline]
-    pub fn elt_size(&self) -> u32 {
-        self.data_room_size as u32
+    pub fn elt_size(&self) -> u16 {
+        self.data_room_size
     }
 }
